docs: add Globals component documentation
This commit adds the documentation for the Globals component that allows the use of global components provided by naive-ui in the entire application. It also includes examples of how to use `$message` and `$dialog` with promises and `await`. Additionally, it shows how to mount only the necessary components using `n-install-provider`.