feat: add support for Naive UI Pro Components, refactor enhancing app logic
Add support for Naive UI Pro Components by including `naive-ui-pro-components` in the list of modules that should not be treated as external by the Vite SSR plugin. Refactor the `enhanceApp` function to conditionally import `naive-ui` and `naive-ui-pro-components` only on the client side to fix issues caused by accessing the `global` and `document` objects when building the server side bundle.