feat(config): iife format and remove iifeMin format
This commit adds support for the iife format and removes the iifeMin format from the build options. The iife format is now included by default, allowing users to build their package as an immediately invoked function expression (IIFE). The iifeMin format has been removed as it was redundant with the minify flag now enabled by default on the iife format. The esbuild plugin configuration has updated to include the unplugin-vue-jsx plugin in both the default and iife formats.