docs(form): add array demo
Added a new demo file to show how `defineForm` can be used in array mode to merge fields within an array. This allows for more flexibility in defining forms with complex data structures. The demo also includes a simple example of using a reactive reference for a form field value.