feat(components): add expand prop to NaiveUIContainer and Checkbox demos
Previously, the NaiveUIContainer and Checkbox demos only had fixed sizes. With this commit, an "expand" prop has been added to the demos to allow them to be expanded to fit the full width of the screen if desired.

Additionally, the "expand" prop has been enabled for the "基础用法" demo of the Checkbox component.