feat(radio): add support for button type in ProRadioGroup
Added support a new type 'button' in the ProRadioGroup component. The new type allows users to render radio buttons as buttons instead of default radio inputs. To achieve this, the ProRadio component was modified to render a button instead of a radio input when the 'button' type is passed as a prop. The ProRadio was modified to account for this change and render radio options as button inputs if the 'button' type is selected.

This commit also removes the ProRadio component as it is no longer needed since its only purpose was to differentiate between default and button types, which can now be achieved directly in the modifiedGroup component.