feat: Add global for dialog, notification
This commit adds new global components for dialog, loading bar, message, and notification by leveraging the corresponding components from `naive-ui` library. The added components are `NGlobalDialog`, `NGlobalLoadingBar`, `NGlobalMessage`, and `NGlobalNotification`.

In addition, a new component `NInstallProvider` is added to install the global components. component takes an array of components or component options and resolves them to a list of component options. Finally, resolved component options are passed to `props.providers` of `NGlobalProvider`.